pyproject.toml
[project]
name = "curate-ns-pond"
description = "Pipeline utilities for literature curation of neuroimaging studies"
readme = "README.md"
authors = [
    { name = "James Kent", email = "jamesdkent21@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "typer>=0.12.0",
    "tenacity>=8.2.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.1",
    "beautifulsoup4>=4.12.0"
]
dynamic = ["version"]

[project.optional-dependencies]
fulltext = [
    "lxml>=5.1.0",
    "pubget>=0.0.8",
    "ace @ git+https://github.com/neurosynth/ACE.git"
]
llm = [
    "openai>=1.14.0"
]

[project.scripts]
curate-ns-pond = "curate_ns_pond.cli:app"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/curate_ns_pond/_version.py"

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "black>=24.3.0",
    "isort>=5.13.0",
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.20.2",
    "pytest-cov>=4.1.0",
    "freezegun>=1.4.0",
    "vcrpy>=5.1.0"
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]
python_files = ["test_*.py"]
